Move train speed adaptation structs to separate header file
